Move containers logic to module utils (#139)
